+++ test.js
@@ -2,20 +2,20 @@
 
 'use strict'
 
-var shell = require('shelljs')
-var fs = require('fs')
-var path = require('path')
-var stream = require('stream')
-var mockGit = require('mock-git')
-var mockery = require('mockery')
-var semver = require('semver')
-var formatCommitMessage = require('./lib/format-commit-message')
-var cli = require('./command')
-var standardVersion = require('./index')
+const shell = require('shelljs')
+const fs = require('fs')
+const path = require('path')
+const stream = require('stream')
+const mockGit = require('mock-git')
+const mockery = require('mockery')
+const semver = require('semver')
+const formatCommitMessage = require('./lib/format-commit-message')
+const cli = require('./command')
+const standardVersion = require('./index')
 
 require('chai').should()
 
-var cliPath = path.resolve(__dirname, './bin/cli.js')
+let cliPath = path.resolve(__dirname, './bin/cli.js')
 
 function branch (branch) {
   shell.exec('git branch ' + branch)
@@ -43,32 +43,31 @@ function execCliAsync (argString) {
 
 function writePackageJson (version, option) {
   option = option || {}
-  var pkg = Object.assign(option, {version: version})
+  let pkg = Object.assign(option, { version: version })
   fs.writeFileSync('package.json', JSON.stringify(pkg), 'utf-8')
-  delete require.cache[require.resolve(path.join(process.cwd(), 'package.json'))]
 }
 
 function writeBowerJson (version, option) {
   option = option || {}
-  var bower = Object.assign(option, {version: version})
+  let bower = Object.assign(option, { version: version })
   fs.writeFileSync('bower.json', JSON.stringify(bower), 'utf-8')
 }
 
 function writeManifestJson (version, option) {
   option = option || {}
-  var manifest = Object.assign(option, {version: version})
+  let manifest = Object.assign(option, { version: version })
   fs.writeFileSync('manifest.json', JSON.stringify(manifest), 'utf-8')
 }
 
 function writeNpmShrinkwrapJson (version, option) {
   option = option || {}
-  var shrinkwrap = Object.assign(option, { version: version })
+  let shrinkwrap = Object.assign(option, { version: version })
   fs.writeFileSync('npm-shrinkwrap.json', JSON.stringify(shrinkwrap), 'utf-8')
 }
 
 function writePackageLockJson (version, option) {
   option = option || {}
-  var pkgLock = Object.assign(option, { version: version })
+  let pkgLock = Object.assign(option, { version: version })
   fs.writeFileSync('package-lock.json', JSON.stringify(pkgLock), 'utf-8')
 }
 
@@ -83,7 +82,7 @@ function writePostBumpHook (causeError) {
 
 function writeHook (hookName, causeError, script) {
   shell.mkdir('-p', 'scripts')
-  var content = script || 'console.error("' + hookName + ' ran")'
+  let content = script || 'console.error("' + hookName + ' ran")'
   content += causeError ? '\nthrow new Error("' + hookName + '-failure")' : ''
   fs.writeFileSync('scripts/' + hookName + '.js', content, 'utf-8')
   fs.chmodSync('scripts/' + hookName + '.js', '755')
@@ -132,7 +131,7 @@ describe('cli', function () {
 
       execCli().code.should.equal(0)
 
-      var content = fs.readFileSync('CHANGELOG.md', 'utf-8')
+      let content = fs.readFileSync('CHANGELOG.md', 'utf-8')
       content.should.match(/patch release/)
       content.should.not.match(/first commit/)
     })
@@ -144,7 +143,7 @@ describe('cli', function () {
       commit('fix: patch release')
       execCli('--first-release').code.should.equal(0)
 
-      var content = fs.readFileSync('CHANGELOG.md', 'utf-8')
+      let content = fs.readFileSync('CHANGELOG.md', 'utf-8')
       content.should.match(/patch release/)
       content.should.match(/first commit/)
       shell.exec('git tag').stdout.should.match(/1\.0\.1/)
@@ -160,7 +159,7 @@ describe('cli', function () {
       commit('fix: patch release')
 
       execCli().code.should.equal(0)
-      var content = fs.readFileSync('CHANGELOG.md', 'utf-8')
+      let content = fs.readFileSync('CHANGELOG.md', 'utf-8')
       content.should.match(/1\.0\.1/)
       content.should.not.match(/legacy header format/)
     })
@@ -178,8 +177,8 @@ describe('cli', function () {
 
       execCli('--commit-all').code.should.equal(0)
 
-      var content = fs.readFileSync('CHANGELOG.md', 'utf-8')
-      var status = shell.exec('git status --porcelain') // see http://unix.stackexchange.com/questions/155046/determine-if-git-working-directory-is-clean-from-a-script
+      let content = fs.readFileSync('CHANGELOG.md', 'utf-8')
+      let status = shell.exec('git status --porcelain') // see http://unix.stackexchange.com/questions/155046/determine-if-git-working-directory-is-clean-from-a-script
 
       status.should.equal('')
       status.should.not.match(/STUFF.md/)
@@ -196,7 +195,7 @@ describe('cli', function () {
         .then(function (unmock) {
           execCli('--sign').code.should.equal(0)
 
-          var captured = shell.cat('gitcapture.log').stdout.split('\n').map(function (line) {
+          let captured = shell.cat('gitcapture.log').stdout.split('\n').map(function (line) {
             return line ? JSON.parse(line) : line
           })
           captured[captured.length - 3].should.deep.equal(['commit', '-S', 'CHANGELOG.md', 'package.json', '-m', 'chore(release): 1.0.1'])
@@ -210,7 +209,7 @@ describe('cli', function () {
       // mock git by throwing on attempt to commit
       return mockGit('console.error("commit yourself"); process.exit(128);', 'commit')
         .then(function (unmock) {
-          var result = execCli()
+          let result = execCli()
           result.code.should.equal(1)
           result.stderr.should.match(/commit yourself/)
 
@@ -222,7 +221,7 @@ describe('cli', function () {
       // mock git by throwing on attempt to add
       return mockGit('console.error("addition is hard"); process.exit(128);', 'add')
         .then(function (unmock) {
-          var result = execCli()
+          let result = execCli()
           result.code.should.equal(1)
           result.stderr.should.match(/addition is hard/)
 
@@ -234,7 +233,7 @@ describe('cli', function () {
       // mock git by throwing on attempt to commit
       return mockGit('console.error("tag, you\'re it"); process.exit(128);', 'tag')
         .then(function (unmock) {
-          var result = execCli()
+          let result = execCli()
           result.code.should.equal(1)
           result.stderr.should.match(/tag, you're it/)
 
@@ -248,7 +247,7 @@ describe('cli', function () {
         .then(function (unmock) {
           writePackageJson('1.0.0')
 
-          var result = execCli()
+          let result = execCli()
           result.code.should.equal(0)
           result.stderr.should.match(/haha, kidding, this is just a warning/)
 
@@ -271,7 +270,7 @@ describe('cli', function () {
         fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
 
         commit('feat: first commit')
-        var result = execCli('--patch')
+        let result = execCli('--patch')
         result.code.should.equal(0)
         result.stderr.should.match(/prerelease ran/)
       })
@@ -288,7 +287,7 @@ describe('cli', function () {
         fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
 
         commit('feat: first commit')
-        var result = execCli('--patch')
+        let result = execCli('--patch')
         result.code.should.equal(1)
         result.stderr.should.match(/prerelease ran/)
       })
@@ -307,7 +306,7 @@ describe('cli', function () {
         fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
 
         commit('feat: first commit')
-        var result = execCli('--patch')
+        let result = execCli('--patch')
         result.stdout.should.match(/9\.9\.9/)
         result.code.should.equal(0)
       })
@@ -326,7 +325,7 @@ describe('cli', function () {
         fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
 
         commit('feat: first commit')
-        var result = execCli('--patch')
+        let result = execCli('--patch')
         result.code.should.equal(0)
         result.stderr.should.match(/postbump ran/)
       })
@@ -343,7 +342,7 @@ describe('cli', function () {
         fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
 
         commit('feat: first commit')
-        var result = execCli('--patch')
+        let result = execCli('--patch')
         result.code.should.equal(1)
         result.stderr.should.match(/postbump-failure/)
       })
@@ -362,7 +361,7 @@ describe('cli', function () {
         fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
 
         commit('feat: first commit')
-        var result = execCli('--patch')
+        let result = execCli('--patch')
         result.code.should.equal(0)
         result.stderr.should.match(/precommit ran/)
       })
@@ -379,7 +378,7 @@ describe('cli', function () {
         fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
 
         commit('feat: first commit')
-        var result = execCli('--patch')
+        let result = execCli('--patch')
         result.code.should.equal(1)
         result.stderr.should.match(/precommit-failure/)
       })
@@ -396,7 +395,7 @@ describe('cli', function () {
         fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
 
         commit('feat: first commit')
-        var result = execCli('--patch')
+        let result = execCli('--patch')
         result.code.should.equal(0)
         shell.exec('git log --oneline -n1').should.match(/delivers #222/)
       })
@@ -423,6 +422,22 @@ describe('cli', function () {
       commit('feat: first commit')
       execCli('--prerelease').stdout.should.include('--tag prerelease')
     })
+
+    it('advises use of --tag alpha for publishing to npm when tagging alpha', function () {
+      writePackageJson('1.0.0')
+      fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
+
+      commit('feat: first commit')
+      execCli('--prerelease alpha').stdout.should.include('--tag alpha')
+    })
+
+    it('does not advise use of --tag prerelease for private modules', function () {
+      writePackageJson('1.0.0', { private: true })
+      fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
+
+      commit('feat: first commit')
+      execCli('--prerelease').stdout.should.not.include('--tag prerelease')
+    })
   })
 
   describe('manual-release', function () {
@@ -435,11 +450,11 @@ describe('cli', function () {
     })
 
     describe('release-types', function () {
-      var regularTypes = ['major', 'minor', 'patch']
+      let regularTypes = ['major', 'minor', 'patch']
 
       regularTypes.forEach(function (type) {
         it('creates a ' + type + ' release', function () {
-          var originVer = '1.0.0'
+          let originVer = '1.0.0'
           writePackageJson(originVer)
           fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
 
@@ -447,7 +462,7 @@ describe('cli', function () {
 
           return execCliAsync('--release-as ' + type)
             .then(function () {
-              var version = {
+              let version = {
                 major: semver.major(originVer),
                 minor: semver.minor(originVer),
                 patch: semver.patch(originVer)
@@ -463,7 +478,7 @@ describe('cli', function () {
       // this is for pre-releases
       regularTypes.forEach(function (type) {
         it('creates a pre' + type + ' release', function () {
-          var originVer = '1.0.0'
+          let originVer = '1.0.0'
           writePackageJson(originVer)
           fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
 
@@ -471,7 +486,7 @@ describe('cli', function () {
 
           return execCliAsync('--release-as ' + type + ' --prerelease ' + type)
             .then(function () {
-              var version = {
+              let version = {
                 major: semver.major(originVer),
                 minor: semver.minor(originVer),
                 patch: semver.patch(originVer)
@@ -487,7 +502,7 @@ describe('cli', function () {
 
     describe('release-as-exact', function () {
       it('releases as v100.0.0', function () {
-        var originVer = '1.0.0'
+        let originVer = '1.0.0'
         writePackageJson(originVer)
         fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
 
@@ -500,7 +515,7 @@ describe('cli', function () {
       })
 
       it('releases as 200.0.0-amazing', function () {
-        var originVer = '1.0.0'
+        let originVer = '1.0.0'
         writePackageJson(originVer)
         fs.writeFileSync('CHANGELOG.md', 'legacy header format<a name="1.0.0">\n', 'utf-8')
 
@@ -566,7 +581,7 @@ describe('cli', function () {
 
     execCli().code.should.equal(0)
 
-    var content = fs.readFileSync('CHANGELOG.md', 'utf-8')
+    let content = fs.readFileSync('CHANGELOG.md', 'utf-8')
     content.should.match(/this is my fairly long commit message which is testing whether or not we allow for long commit messages/)
   })
 
@@ -586,8 +601,46 @@ describe('cli', function () {
   it('appends line feed at end of package.json', function () {
     execCli().code.should.equal(0)
 
-    var pkgJson = fs.readFileSync('package.json', 'utf-8')
+    let pkgJson = fs.readFileSync('package.json', 'utf-8')
     pkgJson.should.equal(['{', '  "version": "1.0.1"', '}', ''].join('\n'))
+  })
+
+  it('preserves indentation of tabs in package.json', function () {
+    let indentation = '\t'
+    let newPkgJson = ['{', indentation + '"version": "1.0.0"', '}', ''].join('\n')
+    fs.writeFileSync('package.json', newPkgJson, 'utf-8')
+
+    execCli().code.should.equal(0)
+    let pkgJson = fs.readFileSync('package.json', 'utf-8')
+    pkgJson.should.equal(['{', indentation + '"version": "1.0.1"', '}', ''].join('\n'))
+  })
+
+  it('preserves indentation of spaces in package.json', function () {
+    let indentation = '     '
+    let newPkgJson = ['{', indentation + '"version": "1.0.0"', '}', ''].join('\n')
+    fs.writeFileSync('package.json', newPkgJson, 'utf-8')
+
+    execCli().code.should.equal(0)
+    let pkgJson = fs.readFileSync('package.json', 'utf-8')
+    pkgJson.should.equal(['{', indentation + '"version": "1.0.1"', '}', ''].join('\n'))
+  })
+
+  it('preserves line feed in package.json', function () {
+    let newPkgJson = ['{', '  "version": "1.0.0"', '}', ''].join('\n')
+    fs.writeFileSync('package.json', newPkgJson, 'utf-8')
+
+    execCli().code.should.equal(0)
+    let pkgJson = fs.readFileSync('package.json', 'utf-8')
+    pkgJson.should.equal(['{', '  "version": "1.0.1"', '}', ''].join('\n'))
+  })
+
+  it('preserves carriage return + line feed in package.json', function () {
+    let newPkgJson = ['{', '  "version": "1.0.0"', '}', ''].join('\r\n')
+    fs.writeFileSync('package.json', newPkgJson, 'utf-8')
+
+    execCli().code.should.equal(0)
+    let pkgJson = fs.readFileSync('package.json', 'utf-8')
+    pkgJson.should.equal(['{', '  "version": "1.0.1"', '}', ''].join('\r\n'))
   })
 
   it('does not run git hooks if the --no-verify flag is passed', function () {
@@ -600,22 +653,34 @@ describe('cli', function () {
   })
 
   it('does not print output when the --silent flag is passed', function () {
-    var result = execCli('--silent')
+    let result = execCli('--silent')
     result.code.should.equal(0)
     result.stdout.should.equal('')
     result.stderr.should.equal('')
   })
 
   it('does not display `npm publish` if the package is private', function () {
-    writePackageJson('1.0.0', {private: true})
+    writePackageJson('1.0.0', { private: true })
 
-    var result = execCli()
+    let result = execCli()
     result.code.should.equal(0)
     result.stdout.should.not.match(/npm publish/)
   })
 
+  it('does not display `all staged files` without the --commit-all flag', function () {
+    var result = execCli()
+    result.code.should.equal(0)
+    result.stdout.should.not.match(/and all staged files/)
+  })
+
+  it('does display `all staged files` if the --commit-all flag is passed', function () {
+    var result = execCli('--commit-all')
+    result.code.should.equal(0)
+    result.stdout.should.match(/and all staged files/)
+  })
+
   it('includes merge commits', function () {
-    var branchName = 'new-feature'
+    let branchName = 'new-feature'
     commit('feat: first commit')
     shell.exec('git tag -a v1.0.0 -m "my awesome first release"')
     branch(branchName)
@@ -626,10 +691,10 @@ describe('cli', function () {
 
     execCli().code.should.equal(0)
 
-    var content = fs.readFileSync('CHANGELOG.md', 'utf-8')
+    let content = fs.readFileSync('CHANGELOG.md', 'utf-8')
     content.should.match(/new feature from branch/)
 
-    var pkgJson = fs.readFileSync('package.json', 'utf-8')
+    let pkgJson = fs.readFileSync('package.json', 'utf-8')
     pkgJson.should.equal(['{', '  "version": "1.1.0"', '}', ''].join('\n'))
   })
 
@@ -641,7 +706,7 @@ describe('cli', function () {
     })
 
     commit('feat: first commit')
-    var result = execCli()
+    let result = execCli()
     result.code.should.equal(1)
     result.stderr.should.match(/scripts must be an object/)
   })
@@ -654,7 +719,7 @@ describe('cli', function () {
     })
 
     commit('feat: first commit')
-    var result = execCli()
+    let result = execCli()
     result.code.should.equal(1)
     result.stderr.should.match(/skip must be an object/)
   })
@@ -666,7 +731,7 @@ describe('standard-version', function () {
 
   describe('with mocked conventionalRecommendedBump', function () {
     beforeEach(function () {
-      mockery.enable({warnOnUnregistered: false, useCleanCache: true})
+      mockery.enable({ warnOnUnregistered: false, useCleanCache: true })
       mockery.registerMock('conventional-recommended-bump', function (_, cb) {
         cb(new Error('bump err'))
       })
@@ -682,7 +747,7 @@ describe('standard-version', function () {
       shell.exec('git tag -a v1.0.0 -m "my awesome first release"')
       commit('feat: new feature!')
 
-      require('./index')({silent: true})
+      require('./index')({ silent: true })
         .catch((err) => {
           err.message.should.match(/bump err/)
           done()
@@ -692,9 +757,9 @@ describe('standard-version', function () {
 
   describe('with mocked conventionalChangelog', function () {
     beforeEach(function () {
-      mockery.enable({warnOnUnregistered: false, useCleanCache: true})
+      mockery.enable({ warnOnUnregistered: false, useCleanCache: true })
       mockery.registerMock('conventional-changelog', function () {
-        var readable = new stream.Readable({objectMode: true})
+        let readable = new stream.Readable({ objectMode: true })
         readable._read = function () {
         }
         setImmediate(readable.emit.bind(readable), 'error', new Error('changelog err'))
@@ -712,7 +777,7 @@ describe('standard-version', function () {
       shell.exec('git tag -a v1.0.0 -m "my awesome first release"')
       commit('feat: new feature!')
 
-      require('./index')({silent: true})
+      require('./index')({ silent: true })
         .catch((err) => {
           err.message.should.match(/changelog err/)
           return done()
@@ -725,7 +790,7 @@ describe('standard-version', function () {
     shell.exec('git tag -a v1.0.0 -m "my awesome first release"')
     commit('feat: new feature!')
 
-    require('./index')({silent: true})
+    require('./index')({ silent: true })
       .then(() => {
         // check last commit message
         shell.exec('git log --oneline -n1').stdout.should.match(/chore\(release\): 1\.1\.0/)
@@ -738,7 +803,10 @@ describe('standard-version', function () {
   describe('without a package file to bump', function () {
     it('should exit with error', function () {
       shell.rm('package.json')
-      return require('./index')({silent: true})
+      return require('./index')({
+        silent: true,
+        gitTagFallback: false
+      })
         .catch((err) => {
           err.message.should.equal('no package file found')
         })
@@ -754,7 +822,7 @@ describe('standard-version', function () {
       commit('feat: first commit')
       shell.exec('git tag -a v1.0.0 -m "my awesome first release"')
       commit('feat: new feature!')
-      return require('./index')({silent: true})
+      return require('./index')({ silent: true })
         .then(() => {
           JSON.parse(fs.readFileSync('bower.json', 'utf-8')).version.should.equal('1.1.0')
           getPackageVersion().should.equal('1.1.0')
@@ -771,7 +839,7 @@ describe('standard-version', function () {
       commit('feat: first commit')
       shell.exec('git tag -a v1.0.0 -m "my awesome first release"')
       commit('feat: new feature!')
-      return require('./index')({silent: true})
+      return require('./index')({ silent: true })
         .then(() => {
           JSON.parse(fs.readFileSync('manifest.json', 'utf-8')).version.should.equal('1.1.0')
           getPackageVersion().should.equal('1.1.0')
@@ -788,7 +856,7 @@ describe('standard-version', function () {
       commit('feat: first commit')
       shell.exec('git tag -a v1.0.0 -m "my awesome first release"')
       commit('feat: new feature!')
-      require('./index')({silent: true})
+      require('./index')({ silent: true })
         .then(() => {
           JSON.parse(fs.readFileSync('npm-shrinkwrap.json', 'utf-8')).version.should.equal('1.1.0')
           getPackageVersion().should.equal('1.1.0')
@@ -807,7 +875,7 @@ describe('standard-version', function () {
       commit('feat: first commit')
       shell.exec('git tag -a v1.0.0 -m "my awesome first release"')
       commit('feat: new feature!')
-      return require('./index')({silent: true})
+      return require('./index')({ silent: true })
         .then(() => {
           JSON.parse(fs.readFileSync('package-lock.json', 'utf-8')).version.should.equal('1.1.0')
           getPackageVersion().should.equal('1.1.0')
@@ -838,7 +906,7 @@ describe('standard-version', function () {
       return execCliAsync('--skip.bump true --skip.changelog true')
         .then(function () {
           getPackageVersion().should.equal('1.0.0')
-          var content = fs.readFileSync('CHANGELOG.md', 'utf-8')
+          let content = fs.readFileSync('CHANGELOG.md', 'utf-8')
           content.should.equal(changelogContent)
         })
     })
@@ -852,7 +920,7 @@ describe('standard-version', function () {
       return execCliAsync('--skip.commit true')
         .then(function () {
           getPackageVersion().should.equal('1.1.0')
-          var content = fs.readFileSync('CHANGELOG.md', 'utf-8')
+          let content = fs.readFileSync('CHANGELOG.md', 'utf-8')
           content.should.match(/new feature from branch/)
           // check last commit message
           shell.exec('git log --oneline -n1').stdout.should.match(/feat: new feature from branch/)
@@ -871,10 +939,53 @@ describe('standard-version', function () {
       commit('feat: first commit')
       shell.exec('git tag -a v1.0.0 -m "my awesome first release"')
       commit('feat: new feature!')
-      return require('./index')({silent: true})
+      return require('./index')({ silent: true })
         .then(() => {
           JSON.parse(fs.readFileSync('bower.json', 'utf-8')).version.should.equal('1.0.0')
           getPackageVersion().should.equal('1.1.0')
+        })
+    })
+  })
+
+  describe('.gitignore', () => {
+    beforeEach(function () {
+      writeBowerJson('1.0.0')
+    })
+
+    it('does not update files present in .gitignore', () => {
+      fs.writeFileSync('.gitignore', 'bower.json', 'utf-8')
+
+      commit('feat: first commit')
+      shell.exec('git tag -a v1.0.0 -m "my awesome first release"')
+      commit('feat: new feature!')
+      return require('./index')({ silent: true })
+        .then(() => {
+          JSON.parse(fs.readFileSync('bower.json', 'utf-8')).version.should.equal('1.0.0')
+          getPackageVersion().should.equal('1.1.0')
+        })
+    })
+  })
+
+  describe('gitTagFallback', () => {
+    it('defaults to 1.0.0 if no tags in git history', () => {
+      shell.rm('package.json')
+      commit('feat: first commit')
+      return require('./index')({ silent: true })
+        .then(() => {
+          const output = shell.exec('git tag')
+          output.stdout.should.include('v1.1.0')
+        })
+    })
+
+    it('bases version on last tag, if tags are found', () => {
+      shell.rm('package.json')
+      shell.exec('git tag -a v5.0.0 -m "a release"')
+      shell.exec('git tag -a v3.0.0 -m "another release"')
+      commit('feat: another commit')
+      return require('./index')({ silent: true })
+        .then(() => {
+          const output = shell.exec('git tag')
+          output.stdout.should.include('v5.1.0')
         })
     })
   })